+++ ios/RCTConvert+BitmovinPlayer.swift
@@ -932,4 +932,26 @@ extension RCTConvert {
 
         return adaptationConfig
     }
+
+    /**
+     Utility method to compute a JS value from an `Thumbnail` object.
+     - Parameter thumbnail `Thumbnail` object to be converted.
+     - Returns: The produced JS object.
+     */
+    static func toJson(thumbnail: Thumbnail?) -> [String: Any?]? {
+        guard let thumbnail = thumbnail else {
+            return nil
+        }
+
+        return [
+            "start": thumbnail.start,
+            "end": thumbnail.end,
+            "text": thumbnail.text,
+            "url": thumbnail.url.absoluteString,
+            "x": thumbnail.x,
+            "y": thumbnail.y,
+            "width": thumbnail.width,
+            "height": thumbnail.height,
+        ]
+    }
 }